Added some new classes, networking now depends on messageQueue

+++ stocks/networking/src/test/java/nl/rug/aoop/networking/TestClient.java
@@ -82,6 +82,8 @@ public class TestClient {
 
     @AfterAll
     public static void endServer() {
-        server.terminate();
+        if(server.getService() != null) {
+            server.terminate();
+        }
     }
 }
